Added synchronisation to libraries
I removed a lot of code from the library selection panel. Now that the
taskbar presents the errors of downloads, there is no need for the
selection panel to tell which library was downloaded fine and which
wasn't.

However, there was a real need for synchronisation in the way the list
of libraries were built. This is now fixed.